Update RingBufs API - allow referring individual maps
This commit introduces a change to the way RingBuf maps are handled.

Previously, they were coalesced with PerfEventArray maps so if a user
wanted to handle different maps in different async tasks, they needed
to redirect events to e.g. mpsc queues to handle them separately.

This change simplifies the API and implementation by allowing to refer
to BPF maps individually through a corresponding key in a `HashMap`.

Signed-off-by: Nikita Baksalyar <nikita.baksalyar@gmail.com>